Multiple client wallets (#26)
Heavy refactoring of blockchain tools from chain-based to
interface-based
 - Removed BlockchainManager
 - Added `BlockchainClientInterfaces`
 - Added `BlockchainServerInterfaces`.
 - Moved jotai state to `@cfce/state` package

Server interfaces that manage multiple chains (e.g. Web3Interface, now
named `evm`) use the `setChain` function to change which chain is used.

Client interfaces set the chain with the `connect` function
automatically. `Connect` can be called with a specific chain ID to
connect to that chain, otherwise it will connect to the active one in
metamask.

---------

Co-authored-by: Lawal Abubakar Babatunde <utility4all@gmail.com>